Extract M4 dense QKV Metal runtime API
