+++ crates/bitnet-kernels/src/metal/mod.rs
@@ -4,4 +4,5 @@
 //! BitNet model loading, packed-kernel families, and benchmarks live in later
 //! Apple M4 campaign items.
 
+pub mod dense_prefill_qkv;
 pub mod smoke;
