+++ docs/tracking/campaigns/apple-m4-slm-metal-phases/CAMPAIGN.md
@@ -27,7 +27,7 @@ otherwise.
 | M4-METAL-002 | merged | Added the env-gated CPU/Metal Q/K/V parity fixture. |
 | M4-METAL-003 | merged | Integrated phase receipt validation. |
 | M4-METAL-004 | merged | Recorded the resident-route runtime boundary and prerequisite path. |
-| M4-METAL-005 | ready | Promote Q/K/V Metal dispatch from test-only fixture to runtime API. |
+| M4-METAL-005 | in progress | Promote Q/K/V Metal dispatch from test-only fixture to runtime API. |
 | M4-METAL-006 | blocked | Route the phase in resident sessions with parity. |
 | M4-METAL-007 | blocked | Record measured phase-local timing deltas. |
